+++ book/snippets/python/src/keys.py
@@ -22,12 +22,3 @@ def restore():
     secret_key = SecretKey.parse("6b911fd37cdf5c81d4c0adb1ab7fa822ed253ab0ad9aa18d77257c88b29b718e")
     keys = Keys(secret_key)
 # ANCHOR_END: restore
-
-print()
-# ANCHOR: vanity
-def vanity():
-    keys = Keys.vanity(["yuk0"], True, 8)
-    print(" Vanity:")
-    print(f"     Public keys: {keys.public_key().to_bech32()}")
-    print(f"     Secret keys: {keys.secret_key().to_bech32()}")
-# ANCHOR_END: vanity