+++ book/snippets/rust/src/keys.rs
@@ -40,12 +40,3 @@ pub fn restore() -> Result<()> {
     Ok(())
 }
 // ANCHOR_END: restore
-
-// ANCHOR: vanity
-pub fn vanity() -> Result<()> {
-    let keys = Keys::vanity(vec!["0000", "yuk", "yuk0"], true, 8)?;
-    println!("Secret key: {}", keys.secret_key().to_bech32()?);
-    println!("Public key: {}", keys.public_key().to_bech32()?);
-    Ok(())
-}
-// ANCHOR_END: vanity